+++ .github/CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 (Dates are in YYYY-MM-DD format. This message is mainly for my own sake.)
 
 ## [Unreleased]
+### Added
+* Added channel link to `#rules-welcome` on verification fail text
 ### Fixed
 * Bios now show even if the associated PluralKit system cannot be found